refactor: Moving ParseFormPostResponse test helper to internal package
