+++ authorize_helper_test.go
@@ -19,13 +19,16 @@
  *
  */
 
-package fosite
+package fosite_test
 
 import (
 	"bytes"
 	"io/ioutil"
 	"net/url"
 	"testing"
+
+	"github.com/ory/fosite"
+	"github.com/ory/fosite/internal"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -45,7 +48,7 @@ func TestIsLocalhost(t *testing.T) {
 		{expect: true, rawurl: "https://test.localhost"},
 	} {
 		u, _ := url.Parse(c.rawurl)
-		assert.Equal(t, c.expect, IsLocalhost(u), "case %d", k)
+		assert.Equal(t, c.expect, fosite.IsLocalhost(u), "case %d", k)
 	}
 }
 
@@ -62,172 +65,172 @@ func TestIsLocalhost(t *testing.T) {
 // of pre-registered redirect URIs (see Section 5.2.3.5).
 func TestDoesClientWhiteListRedirect(t *testing.T) {
 	for k, c := range []struct {
-		client   Client
+		client   fosite.Client
 		url      string
 		isError  bool
 		expected string
 	}{
 		{
-			client:  &DefaultClient{RedirectURIs: []string{""}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{""}},
 			url:     "https://foo.com/cb",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"wta://auth"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"wta://auth"}},
 			url:      "wta://auth",
 			expected: "wta://auth",
 			isError:  false,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"wta:///auth"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"wta:///auth"}},
 			url:      "wta:///auth",
 			expected: "wta:///auth",
 			isError:  false,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"wta://foo/auth"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"wta://foo/auth"}},
 			url:      "wta://foo/auth",
 			expected: "wta://foo/auth",
 			isError:  false,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"https://bar.com/cb"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"https://bar.com/cb"}},
 			url:     "https://foo.com/cb",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"https://bar.com/cb"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"https://bar.com/cb"}},
 			url:      "",
 			isError:  false,
 			expected: "https://bar.com/cb",
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{""}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{""}},
 			url:     "",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"https://bar.com/cb"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"https://bar.com/cb"}},
 			url:      "https://bar.com/cb",
 			isError:  false,
 			expected: "https://bar.com/cb",
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"https://bar.com/cb"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"https://bar.com/cb"}},
 			url:     "https://bar.com/cb123",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"http://[::1]"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"http://[::1]"}},
 			url:      "http://[::1]:1024",
 			expected: "http://[::1]:1024",
 			isError:  false,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://[::1]"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://[::1]"}},
 			url:     "http://[::1]:1024/cb",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"http://[::1]/cb"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"http://[::1]/cb"}},
 			url:      "http://[::1]:1024/cb",
 			expected: "http://[::1]:1024/cb",
 			isError:  false,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://[::1]"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://[::1]"}},
 			url:     "http://foo.bar/bar",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
 			url:      "http://127.0.0.1:1024",
 			expected: "http://127.0.0.1:1024",
 			isError:  false,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"http://127.0.0.1/cb"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1/cb"}},
 			url:      "http://127.0.0.1:64000/cb",
 			expected: "http://127.0.0.1:64000/cb",
 			isError:  false,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
 			url:     "http://127.0.0.1:64000/cb",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
 			url:      "http://127.0.0.1",
 			expected: "http://127.0.0.1",
 			isError:  false,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"http://127.0.0.1/Cb"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1/Cb"}},
 			url:      "http://127.0.0.1:8080/Cb",
 			expected: "http://127.0.0.1:8080/Cb",
 			isError:  false,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
 			url:     "http://foo.bar/bar",
 			isError: true,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1"}},
 			url:     ":/invalid.uri)bar",
 			isError: true,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb"}},
 			url:     "http://127.0.0.1:8080/Cb",
 			isError: true,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb"}},
 			url:     "http://127.0.0.1:8080/cb?foo=bar",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb?foo=bar"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb?foo=bar"}},
 			url:      "http://127.0.0.1:8080/cb?foo=bar",
 			expected: "http://127.0.0.1:8080/cb?foo=bar",
 			isError:  false,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb?foo=bar"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb?foo=bar"}},
 			url:     "http://127.0.0.1:8080/cb?baz=bar&foo=bar",
 			isError: true,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb?foo=bar&baz=bar"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb?foo=bar&baz=bar"}},
 			url:     "http://127.0.0.1:8080/cb?baz=bar&foo=bar",
 			isError: true,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"https://www.ory.sh/cb"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"https://www.ory.sh/cb"}},
 			url:     "http://127.0.0.1:8080/cb",
 			isError: true,
 		},
 		{
-			client:  &DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb"}},
+			client:  &fosite.DefaultClient{RedirectURIs: []string{"http://127.0.0.1:8080/cb"}},
 			url:     "https://www.ory.sh/cb",
 			isError: true,
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"web+application://callback"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"web+application://callback"}},
 			url:      "web+application://callback",
 			isError:  false,
 			expected: "web+application://callback",
 		},
 		{
-			client:   &DefaultClient{RedirectURIs: []string{"https://google.com/?foo=bar%20foo+baz"}},
+			client:   &fosite.DefaultClient{RedirectURIs: []string{"https://google.com/?foo=bar%20foo+baz"}},
 			url:      "https://google.com/?foo=bar%20foo+baz",
 			isError:  false,
 			expected: "https://google.com/?foo=bar%20foo+baz",
 		},
 	} {
-		redir, err := MatchRedirectURIWithClientRedirectURIs(c.url, c.client)
+		redir, err := fosite.MatchRedirectURIWithClientRedirectURIs(c.url, c.client)
 		assert.Equal(t, c.isError, err != nil, "%d: %+v", k, c)
 		if err == nil {
 			require.NotNil(t, redir, "%d", k)
@@ -255,7 +258,7 @@ func TestIsRedirectURISecure(t *testing.T) {
 	} {
 		uu, err := url.Parse(c.u)
 		require.NoError(t, err)
-		assert.Equal(t, !c.err, IsRedirectURISecure(uu), "case %d", d)
+		assert.Equal(t, !c.err, fosite.IsRedirectURISecure(uu), "case %d", d)
 	}
 }
 
@@ -263,8 +266,8 @@ func TestWriteAuthorizeFormPostResponse(t *testing.T) {
 	var responseBuffer bytes.Buffer
 	redirectURL := "https://localhost:8080/cb"
 	parameters := url.Values{"code": {"lshr755nsg39fgur"}, "state": {"924659540232"}}
-	WriteAuthorizeFormPostResponse(redirectURL, parameters, &responseBuffer)
-	code, state, _, _, _, err := ParseFormPostResponse(redirectURL, ioutil.NopCloser(bytes.NewReader(responseBuffer.Bytes())))
+	fosite.WriteAuthorizeFormPostResponse(redirectURL, parameters, &responseBuffer)
+	code, state, _, _, _, err := internal.ParseFormPostResponse(redirectURL, ioutil.NopCloser(bytes.NewReader(responseBuffer.Bytes())))
 	assert.NoError(t, err)
 	assert.Equal(t, parameters.Get("code"), code)
 	assert.Equal(t, parameters.Get("state"), state)
@@ -288,6 +291,6 @@ func TestIsRedirectURISecureStrict(t *testing.T) {
 	} {
 		uu, err := url.Parse(c.u)
 		require.NoError(t, err)
-		assert.Equal(t, !c.err, IsRedirectURISecureStrict(uu), "case %d", d)
+		assert.Equal(t, !c.err, fosite.IsRedirectURISecureStrict(uu), "case %d", d)
 	}
 }
